+++ src/MovieManagement/ViewModels/Rating/CreateReviewViewModel.cs
@@ -1,6 +1,4 @@
-﻿using MovieManagement.Attributes;
-
-namespace MovieManagement.ViewModels.Rating;
+﻿namespace MovieManagement.ViewModels.Rating;
 
 public class CreateReviewViewModel
 {
@@ -34,30 +32,5 @@ public class CreateReviewViewModel
             ResultMessage = ex.Message;
             throw;
         }
-    }
-}
-
-public class CreateReviewModel {
-    [Required(ErrorMessage = "A rating is required")]
-    [Range(1, 10, ErrorMessage = "A rating must be between {1} and {2}")]
-    public int Rating { get; set; }
-
-    [OptionalStringLength(500, MinimumLength = 3, ErrorMessage = "If present, a review must be between {2} and {1} characters")]
-    public string? Review { get; set; }
-}
-
-public class MovieModel
-{
-    public int Id { get; set; }
-    public string Title { get; set; } = default!;
-    public string? PosterUrl { get; set; }
-    public DateOnly? ReleaseDate { get; set; }
-
-    public MovieModel(Movie movie)
-    {
-        Id = movie.Id;
-        Title = movie.Title;
-        PosterUrl = movie.ImageUrl;
-        ReleaseDate = DateOnly.FromDateTime(movie.ReleaseDate);
     }
 }
